# fs_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for querying a high-performance filesystem index.",
        "title": "File System Index API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/dir/stats": {
            "get": {
                "description": "Returns metadata for a single directory at the specified path, optionally including aggregated size and time stats.",
                "summary": "Get directory metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The directory path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include aggregated size and time stats from dir_stats table",
                        "name": "include_stats",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.FileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/file/stats": {
            "get": {
                "description": "Returns metadata for a single file at the specified path.",
                "summary": "Get file metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The file path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.FileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/group/list": {
            "get": {
                "description": "Returns a sorted list of all groups and their total data usage.",
                "summary": "List all groups with usage statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field to sort by (total_size_bytes, file_count, name). Default: total_size_bytes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC, DESC). Default: DESC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to return. Default: 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip. Default: 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.UserStats"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/list": {
            "get": {
                "description": "Returns a list of files and directories within the specified path using high-performance hash lookup internally.",
                "summary": "List directory contents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The directory path to list",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include aggregated stats from dir_stats table for directories",
                        "name": "include_stats",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.FileInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/scans": {
            "get": {
                "description": "Returns fs_indexer and fs_aggregator run history (scan_sessions), newest first.",
                "summary": "List scan sessions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items to return. Default: 50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip. Default: 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ScanSession"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/search": {
            "post": {
                "description": "Search files and directories using AND/OR conditions on path, uid, gid, file_type and size, with regex support and sortable output.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Advanced filesystem search",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.FileInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "description": "Returns a sorted list of all users and their total data usage.",
                "summary": "List all users with usage statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field to sort by (total_size_bytes, file_count, name). Default: total_size_bytes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC, DESC). Default: DESC",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to return. Default: 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip. Default: 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.UserStats"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.DirAggregates": {
            "type": "object",
            "properties": {
                "atime_first": {
                    "type": "integer"
                },
                "atime_last": {
                    "type": "integer"
                },
                "ctime_first": {
                    "type": "integer"
                },
                "ctime_last": {
                    "type": "integer"
                },
                "file_count": {
                    "type": "integer"
                },
                "mtime_first": {
                    "type": "integer"
                },
                "mtime_last": {
                    "type": "integer"
                },
                "total_size_bytes": {
                    "type": "integer"
                }
            }
        },
        "main.FileInfo": {
            "type": "object",
            "properties": {
                "aggregates": {
                    "$ref": "#/definitions/main.DirAggregates"
                },
                "atime": {
                    "type": "integer"
                },
                "ctime": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "integer"
                },
                "gid": {
                    "type": "integer"
                },
                "group": {
                    "type": "string"
                },
                "mtime": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "permissions": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "uid": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "main.ScanSession": {
            "type": "object",
            "properties": {
                "ended_at": {
                    "type": "integer"
                },
                "files_scanned": {
                    "type": "integer"
                },
                "scan_type": {
                    "description": "indexer | aggregator",
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "integer"
                },
                "status": {
                    "description": "running | success | failed",
                    "type": "string"
                }
            }
        },
        "main.SearchRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "order": {
                    "description": "ASC | DESC",
                    "type": "string"
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.SearchRule"
                    }
                },
                "sort_by": {
                    "description": "path | size_bytes | uid | gid | file_type | mtime | atime | ctime",
                    "type": "string"
                }
            }
        },
        "main.SearchRule": {
            "type": "object",
            "properties": {
                "connector": {
                    "description": "AND | OR (relative to the previous rule)",
                    "type": "string"
                },
                "field": {
                    "description": "path | uid | gid | file_type",
                    "type": "string"
                },
                "operator": {
                    "description": "contains | equals | starts_with | regex | regex_i | not_equals | gt | lt",
                    "type": "string"
                },
                "value": {
                    "description": "raw value; numeric fields parse this to an int",
                    "type": "string"
                }
            }
        },
        "main.UserStats": {
            "type": "object",
            "properties": {
                "file_count": {
                    "type": "integer"
                },
                "id_type": {
                    "type": "string"
                },
                "id_value": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_size_bytes": {
                    "type": "integer"
                }
            }
        }
    }
}